requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
plotly>=5.19.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fpdf2>=2.7.9
numpy>=1.26.0
